Move admixes to dedicated dir
